Chap03/CustomVision/CustomVision/requirements.txt
absl-py==0.10.0
aiohttp==3.7.3
appdirs==1.4.4
astor==0.7.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.1.0
audioread==2.1.9
azure-ai-textanalytics==5.0.0
azure-cognitiveservices-speech==1.14.0
azure-common==1.1.25
azure-core==1.8.0
azure-iothub-device-client==1.4.6
azure-iothub-service-client==1.4.6
azure-nspkg==3.0.2
azure-storage==0.36.0
backcall==0.1.0
bert4keras==0.10.7
bleach==1.5.0
blinker==1.4
branca==0.3.1
Brotli==1.0.9
cachetools==4.0.0
cdata==0.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
chart-studio==1.0.0
ckiptagger==0.2.1
Click==7.0
colorama==0.4.1
colorlover==0.3.0
combo==0.1.0
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.23
-e git+https://github.com/thtrieu/darkflow.git@eb7e830393f24233032b8578737141528be01d65#egg=darkflow
dataclasses==0.7
DateTime==4.3
decorator==4.4.0
deepmachine==0.7.5
defusedxml==0.6.0
dill==0.3.1.1
Django==1.11.29
django-mongodb-engine==0.6.0
djangorestframework==3.11.1
djangotoolbox==1.8.0
dlib==19.21.0
easyocr==1.1.3
editdistance==0.5.3
elasticsearch==7.11.0
entrypoints==0.3
enum34==1.1.10
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.5.0
Faker==3.0.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
folium==0.10.1
fsspec==0.8.5
future==0.18.2
gast==0.3.3
gdown==3.10.1
genibabel==1.0.0
gensim==4.1.0
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.51.0
graphviz==0.8.4
grpcio==1.35.0
h11==0.10.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
html5lib==0.9999999
huggingface-hub==0.0.17
hyperframe==5.2.0
idna==2.7
idna-ssl==1.1.0
imageio==2.5.0
imbalanced-learn==0.6.2
imblearn==0.0
importlib-metadata==1.5.0
imutils==0.5.3
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.13.3
jieba==0.39
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kaitaistruct==0.8
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-tuner==1.0.2
kiwisolver==1.1.0
ldap3==2.7
librosa==0.8.0
lightgbm==2.3.1
llvmlite==0.31.0
lxml==4.5.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.3.2
mistune==0.8.4
mitmproxy==5.2
mock==3.0.5
mpmath==1.1.0
msrest==0.6.18
mtcnn==0.1.0
multidict==5.1.0
mxnet==1.3.1
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nibabel==3.2.1
nltk==3.4.1
notebook==5.7.8
numba==0.48.0
numpy==1.18.5
oauthlib==3.1.0
onnx==1.6.0
openalpr==1.0.12
OpenCC==1.1.1
opencv-contrib-python==4.4.0.44
opencv-python==4.1.0.25
opt-einsum==3.1.0
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.2
parso==0.4.0
passlib==1.7.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.1.2
plotly==2.7.0
pooch==1.3.0
progressbar2==3.53.1
prometheus-client==0.6.0
promise==2.3
prompt-toolkit==2.0.9
protobuf==3.11.3
publicsuffix2==2.20191221
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.20
pydivert==2.1.0
pydot==1.4.1
Pygments==2.4.0
pymongo==3.11.0
pyod==0.7.8.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
pyrsistent==0.15.2
PySocks==1.7.1
PySoundFile==0.9.0.post1
pytesseract==0.3.4
python-dateutil==2.8.0
python-utils==2.5.6
pytorch-lightning==1.1.6
pytz==2019.1
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.4
regex==2021.9.24
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
resnet==0.1
retrying==1.3.3
rsa==4.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sacremoses==0.0.46
scikit-image==0.17.2
scikit-learn==0.24.0
scikit-optimize==0.5.2
scipy==1.4.1
seaborn==0.11.1
Send2Trash==1.5.0
sentencepiece==0.1.95
six==1.12.0
sklearn==0.0
smart-open==5.2.1
sortedcontainers==2.1.0
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
statsmodels==0.12.1
suod==0.0.3
sympy==1.5.1
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.3.0
tensorflow-datasets==2.1.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
tensorflow-gpu-estimator==2.3.0
tensorflow-metadata==0.21.1
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.2
terminaltables==3.1.0
testpath==0.4.2
text-unidecode==1.3
textblob==0.15.3
threadpoolctl==2.1.0
tifffile==2020.9.3
tokenizers==0.10.3
torch==1.7.1+cpu
torch-lr-finder==0.2.1
torchaudio==0.7.2
torchsummary==1.5.1
torchtext==0.6.0
torchvision==0.8.2+cpu
tornado==6.0.2
tqdm==4.43.0
traitlets==4.3.2
transformers==4.10.3
typing==3.7.4.1
typing-extensions==3.7.4.1
urllib3==1.24.1
urwid==2.0.1
utils==1.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.4.2
wordcloud==1.8.1
wrapt==1.12.0
wsproto==0.15.0
xgboost==0.90
xlrd==1.2.0
xlwt==1.3.0
yarl==1.6.3
zipp==3.0.0
zope.interface==4.6.0
zstandard==0.13.0
